+++ concurrency/Control/Concurrent/Classy/CRef.hs
@@ -1,19 +1,18 @@
 -- |
 -- Module      : Control.Concurrent.Classy.CRef
--- Copyright   : (c) 2016 Michael Walker
+-- Copyright   : (c) 2016--2018 Michael Walker
 -- License     : MIT
 -- Maintainer  : Michael Walker <mike@barrucadu.co.uk>
--- Stability   : stable
+-- Stability   : experimental
 -- Portability : portable
 --
--- Mutable references in a concurrency monad.
---
--- __Deviations:__ There is no @Eq@ instance for @MonadConc@ the
--- @CRef@ type. Furthermore, the @mkWeakIORef@ function is not
--- provided.
-module Control.Concurrent.Classy.CRef
+-- Deprecated re-exports of @IORef@ functions under the old @CRef@
+-- names.
+module Control.Concurrent.Classy.CRef {-# DEPRECATED "Import Control.Concurrent.Classy.IORef instead" #-}
   ( -- * CRefs
-    newCRef
+    CRef
+  , newCRef
+  , newCRefN
   , readCRef
   , writeCRef
   , modifyCRef
@@ -21,6 +20,11 @@ module Control.Concurrent.Classy.CRef
   , atomicModifyCRef
   , atomicModifyCRef'
   , atomicWriteCRef
+
+  -- ** Compare-and-swap
+  , casCRef
+  , modifyCRefCAS
+  , modifyCRefCAS_
 
   -- * Memory Model
 
@@ -80,41 +84,85 @@ module Control.Concurrent.Classy.CRef
   -- memory barrier.
   ) where
 
-import           Control.Monad.Conc.Class
+import qualified Control.Concurrent.Classy.IORef as IORef
+import           Control.Monad.Conc.Class        (IORef, MonadConc, Ticket)
+import qualified Control.Monad.Conc.Class        as IORef
+
+-- | Type alias for 'IORef'.
+type CRef m a = IORef m a
+{-# DEPRECATED CRef "Use IORef instead" #-}
+
+-- | Create a new reference.
+newCRef :: MonadConc m => a -> m (CRef m a)
+newCRef = IORef.newIORef
+{-# DEPRECATED newCRef "Use newIORef instead" #-}
+
+-- | Create a new reference, but it is given a name which may be used
+-- to present more useful debugging information.
+newCRefN :: MonadConc m => String -> a -> m (CRef m a)
+newCRefN = IORef.newIORefN
+{-# DEPRECATED newCRefN "Use newIORefN instead" #-}
+
+-- | Read the current value stored in a reference.
+readCRef :: MonadConc m => CRef m a -> m a
+readCRef = IORef.readIORef
+{-# DEPRECATED readCRef "Use readIORef instead" #-}
+
+-- | Write a new value into an @CRef@, without imposing a memory
+-- barrier. This means that relaxed memory effects can be observed.
+writeCRef :: MonadConc m => CRef m a -> a -> m ()
+writeCRef = IORef.writeIORef
+{-# DEPRECATED writeCRef "Use writeIORef instead" #-}
 
 -- | Mutate the contents of a @CRef@.
 --
 -- Be warned that 'modifyCRef' does not apply the function strictly.
 -- This means if the program calls 'modifyCRef' many times, but
 -- seldomly uses the value, thunks will pile up in memory resulting in
--- a space leak. This is a common mistake made when using a @CRef@ as
--- a counter. For example, the following will likely produce a stack
--- overflow:
---
--- >ref <- newCRef 0
--- >replicateM_ 1000000 $ modifyCRef ref (+1)
--- >readCRef ref >>= print
---
--- To avoid this problem, use 'modifyCRef'' instead.
---
--- @since 1.0.0.0
+-- a space leak.
 modifyCRef :: MonadConc m => CRef m a -> (a -> a) -> m ()
-modifyCRef ref f = readCRef ref >>= writeCRef ref . f
+modifyCRef = IORef.modifyIORef
+{-# DEPRECATED modifyCRef "Use modifyIORef instead" #-}
 
 -- | Strict version of 'modifyCRef'
---
--- @since 1.0.0.0
 modifyCRef' :: MonadConc m => CRef m a -> (a -> a) -> m ()
-modifyCRef' ref f = do
-  x <- readCRef ref
-  writeCRef ref $! f x
+modifyCRef' = IORef.modifyIORef'
+{-# DEPRECATED modifyCRef' "Use modifyIORef' instead" #-}
+
+-- | Atomically modify the value stored in a reference. This imposes
+-- a full memory barrier.
+atomicModifyCRef :: MonadConc m => CRef m a -> (a -> (a, b)) -> m b
+atomicModifyCRef = IORef.atomicModifyIORef
+{-# DEPRECATED atomicModifyCRef "Use atomicModifyIORef instead" #-}
 
 -- | Strict version of 'atomicModifyCRef'. This forces both the value
 -- stored in the @CRef@ as well as the value returned.
---
--- @since 1.0.0.0
 atomicModifyCRef' :: MonadConc m => CRef m a -> (a -> (a,b)) -> m b
-atomicModifyCRef' ref f = do
-  b <- atomicModifyCRef ref $ \a -> case f a of
-    v@(a',_) -> a' `seq` v
-  pure $! b
+atomicModifyCRef' = IORef.atomicModifyIORef'
+{-# DEPRECATED atomicModifyCRef' "Use atomicModifyIORef' instead" #-}
+
+-- | Replace the value stored in a reference, with the
+-- barrier-to-reordering property that 'atomicModifyIORef' has.
+atomicWriteCRef :: MonadConc m => CRef m a -> a -> m ()
+atomicWriteCRef = IORef.atomicWriteIORef
+{-# DEPRECATED atomicWriteCRef "Use atomicWriteIORef instead" #-}
+
+-- | Perform a machine-level compare-and-swap (CAS) operation on a
+-- @CRef@. Returns an indication of success and a @Ticket@ for the
+-- most current value in the @CRef@.
+-- This is strict in the \"new\" value argument.
+casCRef :: MonadConc m => CRef m a -> Ticket m a -> a -> m (Bool, Ticket m a)
+casCRef = IORef.casIORef
+{-# DEPRECATED casCRef "Use casIORef instead" #-}
+
+-- | A replacement for 'atomicModifyCRef' using a compare-and-swap.
+--
+-- This is strict in the \"new\" value argument.
+modifyCRefCAS :: MonadConc m => CRef m a -> (a -> (a, b)) -> m b
+modifyCRefCAS = IORef.modifyIORefCAS
+{-# DEPRECATED modifyCRefCAS "Use modifyIORefCAS instead" #-}
+
+-- | A variant of 'modifyCRefCAS' which doesn't return a result.
+modifyCRefCAS_ :: MonadConc m => CRef m a -> (a -> a) -> m ()
+modifyCRefCAS_ = IORef.modifyIORefCAS_
+{-# DEPRECATED modifyCRefCAS_ "Use modifyIORefCAS_ instead" #-}